Move tags related utility functions to `pkg/util` package.
This patch moves all the logic that handles comparing and updating an
SFN resource tags to a seperate utility package. All functionality can
be reused for SFN StateMachine resource.
