+++ pkg/resource/activity/hooks.go
@@ -18,10 +18,9 @@ import (
 
 	ackcompare "github.com/aws-controllers-k8s/runtime/pkg/compare"
 	ackrtlog "github.com/aws-controllers-k8s/runtime/pkg/runtime/log"
-	ackutil "github.com/aws-controllers-k8s/runtime/pkg/util"
-	svcsdk "github.com/aws/aws-sdk-go/service/sfn"
 
 	svcapitypes "github.com/aws-controllers-k8s/sfn-controller/apis/v1alpha1"
+	commonutil "github.com/aws-controllers-k8s/sfn-controller/pkg/util"
 )
 
 // setResourceAdditionalFields queries and adds the tags to an Activity resource
@@ -31,37 +30,21 @@ func (rm *resourceManager) setResourceAdditionalFields(
 ) (err error) {
 	rlog := ackrtlog.FromContext(ctx)
 	exit := rlog.Trace("rm.setResourceAdditionalFields")
-	defer exit(err)
+	defer func() {
+		exit(err)
+	}()
 
 	// Set activity tags
-	ko.Spec.Tags, err = rm.getActivityTags(ctx, string(*ko.Status.ACKResourceMetadata.ARN))
+	ko.Spec.Tags, err = commonutil.GetResourceTags(
+		ctx,
+		rm.sdkapi,
+		rm.metrics,
+		string(*ko.Status.ACKResourceMetadata.ARN),
+	)
 	if err != nil {
 		return err
 	}
-
 	return nil
-}
-
-// getActivityTags retrieves a resource list of tags.
-func (rm *resourceManager) getActivityTags(ctx context.Context, resourceARN string) ([]*svcapitypes.Tag, error) {
-	listTagsForResourceResponse, err := rm.sdkapi.ListTagsForResourceWithContext(
-		ctx,
-		&svcsdk.ListTagsForResourceInput{
-			ResourceArn: &resourceARN,
-		},
-	)
-	rm.metrics.RecordAPICall("GET", "ListTagsForResource", err)
-	if err != nil {
-		return nil, err
-	}
-	tags := make([]*svcapitypes.Tag, 0, len(listTagsForResourceResponse.Tags))
-	for _, tag := range listTagsForResourceResponse.Tags {
-		tags = append(tags, &svcapitypes.Tag{
-			Key:   tag.Key,
-			Value: tag.Value,
-		})
-	}
-	return tags, nil
 }
 
 // customUpdateActivity patches each of the resource properties in the backend AWS
@@ -73,7 +56,14 @@ func (rm *resourceManager) customUpdateActivity(
 	delta *ackcompare.Delta,
 ) (*resource, error) {
 	if delta.DifferentAt("Spec.Tags") {
-		err := rm.updateActivityTags(ctx, latest, desired)
+		err := commonutil.SyncResourceTags(
+			ctx,
+			rm.sdkapi,
+			rm.metrics,
+			string(*desired.ko.Status.ACKResourceMetadata.ARN),
+			latest.ko.Spec.Tags,
+			desired.ko.Spec.Tags,
+		)
 		if err != nil {
 			return nil, err
 		}
@@ -81,49 +71,6 @@ func (rm *resourceManager) customUpdateActivity(
 	return desired, nil
 }
 
-// updateActivityTags uses TagResource and UntagResource to add, remove and update
-// activitytags.
-func (rm *resourceManager) updateActivityTags(
-	ctx context.Context,
-	latest *resource,
-	desired *resource,
-) error {
-	var err error
-	rlog := ackrtlog.FromContext(ctx)
-	exit := rlog.Trace("rm.updateRuleGroupsTags")
-	defer exit(err)
-
-	addedOrUpdated, removed := computeTagsDelta(latest.ko.Spec.Tags, desired.ko.Spec.Tags)
-
-	if len(removed) > 0 {
-		_, err = rm.sdkapi.UntagResourceWithContext(
-			ctx,
-			&svcsdk.UntagResourceInput{
-				ResourceArn: (*string)(latest.ko.Status.ACKResourceMetadata.ARN),
-				TagKeys:     removed,
-			},
-		)
-		rm.metrics.RecordAPICall("UPDATE", "UntagResource", err)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(addedOrUpdated) > 0 {
-		_, err = rm.sdkapi.TagResourceWithContext(
-			ctx,
-			&svcsdk.TagResourceInput{
-				ResourceArn: (*string)(latest.ko.Status.ACKResourceMetadata.ARN),
-				Tags:        sdkTagsFromResourceTags(addedOrUpdated),
-			},
-		)
-		rm.metrics.RecordAPICall("UPDATE", "TagResource", err)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
 func customPreCompare(
 	delta *ackcompare.Delta,
 	a *resource,
@@ -132,66 +79,8 @@ func customPreCompare(
 	if len(a.ko.Spec.Tags) != len(b.ko.Spec.Tags) {
 		delta.Add("Spec.Tags", a.ko.Spec.Tags, b.ko.Spec.Tags)
 	} else if len(a.ko.Spec.Tags) > 0 {
-		if !equalTags(a.ko.Spec.Tags, b.ko.Spec.Tags) {
+		if !commonutil.EqualTags(a.ko.Spec.Tags, b.ko.Spec.Tags) {
 			delta.Add("Spec.Tags", a.ko.Spec.Tags, b.ko.Spec.Tags)
 		}
 	}
-}
-
-// equalTags returns true if two Tag arrays are equal regardless of the order
-// of their elements.
-func equalTags(
-	a []*svcapitypes.Tag,
-	b []*svcapitypes.Tag,
-) bool {
-	addedOrUpdated, removed := computeTagsDelta(a, b)
-	return len(addedOrUpdated) == 0 && len(removed) == 0
-}
-
-// computeTagsDelta compares two Tag arrays and return two different list
-// containing the addedOrupdated and removed tags.
-// The removed tags only contains the Key of tags
-func computeTagsDelta(
-	a []*svcapitypes.Tag,
-	b []*svcapitypes.Tag,
-) (addedOrUpdated []*svcapitypes.Tag, removed []*string) {
-	var visitedIndexes []string
-mainLoop:
-	for _, aElement := range a {
-		visitedIndexes = append(visitedIndexes, *aElement.Key)
-		for _, bElement := range b {
-			if equalStrings(aElement.Key, bElement.Key) {
-				if !equalStrings(aElement.Value, bElement.Value) {
-					addedOrUpdated = append(addedOrUpdated, bElement)
-				}
-				continue mainLoop
-			}
-		}
-		removed = append(removed, aElement.Key)
-	}
-	for _, bElement := range b {
-		if !ackutil.InStrings(*bElement.Key, visitedIndexes) {
-			addedOrUpdated = append(addedOrUpdated, bElement)
-		}
-	}
-	return addedOrUpdated, removed
-}
-
-// svcTagsFromResourceTags transforms a *svcapitypes.Tag array to a *svcsdk.Tag array.
-func sdkTagsFromResourceTags(rTags []*svcapitypes.Tag) []*svcsdk.Tag {
-	tags := make([]*svcsdk.Tag, len(rTags))
-	for i := range rTags {
-		tags[i] = &svcsdk.Tag{
-			Key:   rTags[i].Key,
-			Value: rTags[i].Value,
-		}
-	}
-	return tags
-}
-
-func equalStrings(a, b *string) bool {
-	if a == nil {
-		return b == nil || *b == ""
-	}
-	return (*a == "" && b == nil) || *a == *b
 }
